refactor: flatten internal package layout
